# Optimize CLI startup performance with lazy imports