--- a/yourbench/main.py
+++ b/yourbench/main.py
@@ -8,47 +8,56 @@ from typing import Optional
 from pathlib import Path
 from dataclasses import field, dataclass
 
+
 # Early startup logging
 print("🚀 YourBench starting up...", flush=True)
 startup_time = time.perf_counter()
 
-import yaml
-import typer
-from dotenv import load_dotenv
-from loguru import logger
-from rich.table import Table
-from rich.prompt import Prompt, Confirm, IntPrompt, FloatPrompt
-from rich.console import Console
+import yaml  # noqa: E402
+import typer  # noqa: E402
+from dotenv import load_dotenv  # noqa: E402
+from loguru import logger  # noqa: E402
+from rich.table import Table  # noqa: E402
+from rich.prompt import Prompt, Confirm, IntPrompt, FloatPrompt  # noqa: E402
+from rich.console import Console  # noqa: E402
+
 
 print("⏳ Loading core modules...", flush=True)
 
 # Lazy imports - only import when needed
 launch_ui = None
-run_analysis = None 
+run_analysis = None
 run_pipeline = None
+
 
 def _lazy_import_ui():
     global launch_ui
     if launch_ui is None:
         print("⏳ Loading Gradio UI components...", flush=True)
         from yourbench.app import launch_ui as _launch_ui
+
         launch_ui = _launch_ui
     return launch_ui
+
 
 def _lazy_import_analysis():
     global run_analysis
     if run_analysis is None:
         from yourbench.analysis import run_analysis as _run_analysis
+
         run_analysis = _run_analysis
     return run_analysis
+
 
 def _lazy_import_pipeline():
     global run_pipeline
     if run_pipeline is None:
         print("⏳ Loading pipeline components...", flush=True)
         from yourbench.pipeline.handler import run_pipeline as _run_pipeline
+
         run_pipeline = _run_pipeline
     return run_pipeline
+
 
 print("⏳ Loading environment variables...", flush=True)
 load_dotenv()