--- a/yourbench/utils/configuration_engine.py
+++ b/yourbench/utils/configuration_engine.py
@@ -708,11 +708,13 @@ class YourbenchConfig(BaseModel):
         hf_config_data = data.get("hf_configuration", {})
         if hf_config_data:
             hf_config_data = {k: v for k, v in hf_config_data.items() if v is not None}
-        
+
         config_data = {
             "hf_configuration": HuggingFaceConfig(**hf_config_data),
             "pipeline_config": pipeline_config,
-            "model_list": [ModelConfig(**{k: v for k, v in m.items() if v is not None}) for m in data.get("model_list", [])],
+            "model_list": [
+                ModelConfig(**{k: v for k, v in m.items() if v is not None}) for m in data.get("model_list", [])
+            ],
             "model_roles": data.get("model_roles", {}),
             "debug": data.get("debug", False),
         }